+++ packages/ice-npm-utils/lib/index.js
@@ -52,6 +52,10 @@ function getAndExtractTarball(destDir, tarball, progressFunc = () => {}) {
       .pipe(zlib.Unzip())
       .pipe(new tar.Parse())
       .on('entry', (entry) => {
+        if (entry.type === 'Directory') {
+          entry.resume();
+          return;
+        }
         const realPath = entry.path.replace(/^package\//, '');
 
         let filename = path.basename(realPath);
@@ -103,11 +107,8 @@ function getNpmInfo(npm) {
 
   const register = getNpmRegistry(npm);
   const url = `${register}/${npm}`;
-  log.verbose('getNpmInfo start', url);
 
   return request.get(url).then((response) => {
-    log.verbose('getNpmInfo success');
-
     let body;
     try {
       body = JSON.parse(response);
@@ -170,7 +171,6 @@ function getLatestVersion(npm) {
     }
 
     const latestVersion = data['dist-tags'].latest;
-    log.verbose('getLatestVersion result', npm, latestVersion);
     return latestVersion;
   });
 }
@@ -188,7 +188,6 @@ function getNpmRegistry(npmName = '') {
     return 'https://registry.npm.alibaba-inc.com';
   }
 
-  // TODO: maybe default should be: registry.npm.com
   return 'https://registry.npm.taobao.org';
 }