+++ packages/iceworks-client/src/hooks/useDependency.js
@@ -8,7 +8,7 @@ import stores from '@stores';
 
 export const STATUS_RESETING = 'reseting';
 
-function useDependency(diableUseSocket) {
+function useDependency(diableUseSocket, showGlobalTerminal = true) {
   const {
     on: onCreateModal,
     setModal: setCreateModal,
@@ -68,11 +68,15 @@ function useDependency(diableUseSocket) {
     await dependenciesStore.reset();
 
     setResetModal(false);
-    globalTerminalStore.show();
+
+    if (showGlobalTerminal) {
+      globalTerminalStore.show();
+    }
   }
 
+  // HACK useDependency will be called many times but we don't want to repeat listening
   if (!diableUseSocket) {
-    useSocket('adapter.dependency.reset.data', writeGlobalLog);
+    useSocket('adapter.dependency.reset.data', ({ chunk, isStdout }) => writeGlobalLog(chunk, isStdout));
 
     useSocket('adapter.dependency.reset.exit', (code) => {
       if (code === 0) {
@@ -83,7 +87,7 @@ function useDependency(diableUseSocket) {
       }
     });
 
-    useSocket('adapter.dependency.upgrade.data', writeGlobalLog);
+    useSocket('adapter.dependency.upgrade.data', ({ chunk, isStdout }) => writeGlobalLog(chunk, isStdout));
 
     useSocket('adapter.dependency.upgrade.exit', (code) => {
       if (code === 0) {
@@ -94,7 +98,7 @@ function useDependency(diableUseSocket) {
       }
     });
 
-    useSocket('adapter.dependency.install.data', writeGlobalLog);
+    useSocket('adapter.dependency.install.data', ({ chunk, isStdout }) => writeGlobalLog(chunk, isStdout));
 
     useSocket('adapter.dependency.install.exit', (code) => {
       if (code === 0) {