+++ packages/iceworks-client/src/pages/Project/components/TaskItem/index.js
@@ -2,16 +2,35 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { withRouter } from 'react-router-dom';
 import { FormattedMessage } from 'react-intl';
-import { Button } from '@alifd/next';
+// import { Button } from '@alifd/next';
 import Icon from '@components/Icon';
 import TaskButton from '@components/TaskButton';
+import Modal from '@components/Modal';
 import Panel from '../Panel';
 import PanelHead from '../Panel/head';
 import styles from './index.module.scss';
 
-const TaskItem = ({ title, desc, icon, path, history, isWorking, onStart, onStop }) => {
-  function handleClick() {
+const TaskItem = ({
+  title,
+  desc,
+  icon,
+  path,
+  history,
+  isWorking,
+  onStart,
+  onStop,
+  installDependencyVisible,
+  onInstallDependencyCancel,
+  onInstallDependencyOk,
+}) => {
+  // NOTE: hidden setting button
+  /* function handleClick() {
     history.push(path);
+  } */
+
+  function handleStart() {
+    history.push(path);
+    onStart();
   }
 
   return (
@@ -30,15 +49,23 @@ const TaskItem = ({ title, desc, icon, path, history, isWorking, onStart, onStop
         <div className={styles.taskAction}>
           <TaskButton
             isWorking={isWorking}
-            onStart={onStart}
+            onStart={handleStart}
             onStop={onStop}
           />
-          <Button type="normal" onClick={handleClick} className={styles.settingButton}>
+          {/* <Button type="normal" onClick={handleClick} className={styles.settingButton}>
             <Icon type="settings" className={styles.settingIcon} />
             <FormattedMessage id="iceworks.project.panel.quick.task.setting" />
-          </Button>
+          </Button> */}
         </div>
       </div>
+      <Modal
+        title={<FormattedMessage id="iceworks.project.install.dependencies.title" />}
+        visible={installDependencyVisible}
+        onCancel={onInstallDependencyCancel}
+        onOk={onInstallDependencyOk}
+      >
+        <FormattedMessage id="iceworks.project.install.dependencies.content" />
+      </Modal>
     </Panel>
   );
 };
@@ -56,6 +83,9 @@ TaskItem.propTypes = {
   icon: PropTypes.string.isRequired,
   path: PropTypes.string.isRequired,
   history: PropTypes.object.isRequired,
+  installDependencyVisible: PropTypes.bool.isRequired,
+  onInstallDependencyCancel: PropTypes.func.isRequired,
+  onInstallDependencyOk: PropTypes.func.isRequired,
   isWorking: PropTypes.bool,
   onStart: PropTypes.func,
   onStop: PropTypes.func,