+++ packages/iceworks-client/src/pages/Setting/components/General/index.js
@@ -1,13 +1,13 @@
 /* eslint quote-props:0 */
 import React, { useContext, useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
-import { Grid, Radio } from '@alifd/next';
+import { Grid, Radio, Select, Input } from '@alifd/next';
 import showMessage from '@utils/showMessage';
 import { injectIntl, FormattedMessage } from 'react-intl';
 import { LocalContext, localeInfos } from '@components/Locale';
 import { ThemeContext } from '@components/ThemeProvider';
 import Card from '@components/Card';
-import { THEMES } from '@src/appConfig';
+import appConfig, { THEMES } from '@src/appConfig';
 import socket from '@src/socket';
 import goldlog from '@utils/goldlog';
 import styles from './index.module.scss';
@@ -20,7 +20,9 @@ const General = ({ intl }) => {
   const { locale, setLocale } = useContext(LocalContext);
   const { theme, setTheme } = useContext(ThemeContext);
   const [editor, setEditor] = useState(DEFAULT_EDITOR);
-
+  const [npmClient, setNpmClient] = useState('');
+  const [registry, setRegistry] = useState('');
+  
   // language options
   const languageOptions = Object.keys(localeInfos).map(key => {
     return {
@@ -63,6 +65,28 @@ const General = ({ intl }) => {
     },
   ];
 
+  // npm clients
+  const npmClients = [
+    {
+      label: 'npm',
+      value: 'npm',
+    },
+    {
+      label: 'tnpm',
+      value: 'tnpm',
+    },
+    {
+      label: 'cnpm',
+      value: 'cnpm',
+    },
+    {
+      label: intl.formatMessage({ id: 'iceworks.setting.general.custom.registry' }),
+      value: 'custom',
+    },
+  ].filter(({ value }) => {
+    return value !== 'tnpm' || appConfig.isAliInternal;
+  });
+
   async function onLocaleChange(currentLocale) {
     goldlog({
       namespace: 'home',
@@ -100,12 +124,8 @@ const General = ({ intl }) => {
   }
 
   async function getEditor() {
-    try {
-      const currentLocale = await socket.emit('home.setting.getEditor');
-      setEditor(currentLocale);
-    } catch (error) {
-      showMessage(error);
-    }
+    const currentLocale = await socket.emit('home.setting.getEditor');
+    setEditor(currentLocale);
   }
 
   async function onEditorChange(currentEditor) {
@@ -126,12 +146,54 @@ const General = ({ intl }) => {
     }
   }
 
+  async function getNpmClient() {
+    const currentNpmClient = await socket.emit('home.setting.getNpmClient');
+    setNpmClient(currentNpmClient);
+  }
+
+  async function onClientChange(currentNpmClient) {
+    goldlog({
+      namespace: 'home',
+      module: 'setting',
+      action: 'set-npm-client',
+      data: {
+        npmClient: currentNpmClient,
+      },
+    });
+
+    try {
+      await socket.emit('home.setting.setNpmClient', { npmClient: currentNpmClient });
+      setNpmClient(currentNpmClient);
+    } catch (error) {
+      showMessage(error);
+    }
+  }
+
+  async function getRegistry() {
+    const currentRegistry = await socket.emit('home.setting.getRegistry');
+    setRegistry(currentRegistry);
+  }
+
+  async function onRegistryChange(value) {
+    try {
+      await socket.emit('home.setting.setRegistry', { registry: value });
+      setRegistry(value);
+    } catch (error) {
+      showMessage(error);
+    }
+  }
+
   useEffect(() => {
-    getEditor();
+    try {
+      getEditor();
+      getNpmClient();
+      getRegistry();
+    } catch (error) {
+      showMessage(error);
+    }
   }, []);
 
   return (
-
     <Card title={intl.formatMessage({ id: 'iceworks.setting.general.title' })} contentHeight="100%">
       <Row className={styles.row} key="language">
         <Col span="2" className={styles.label}>
@@ -174,6 +236,28 @@ const General = ({ intl }) => {
           />
         </Col>
       </Row>
+      
+      <Row className={styles.row} key="client">
+        <Col span="2" className={styles.label}>
+          <FormattedMessage id="iceworks.setting.general.npm.client.title" />
+        </Col>
+        <Col span="22">
+          <Select className={styles.select} dataSource={npmClients} value={npmClient} onChange={onClientChange} />
+        </Col>
+      </Row>
+      {npmClient === 'custom' &&
+        <Row className={styles.row} key="cuntom">
+          <Col span="2" className={styles.label} />
+          <Col span="22">
+            <Input
+              onChange={onRegistryChange}
+              style={{ minWidth: 300 }}
+              value={registry}
+              placeholder={intl.formatMessage({ id: 'iceworks.setting.general.custom.placeholder' })}
+            />
+          </Col>
+        </Row>
+      }
     </Card>
   );
 };