+++ packages/iceworks-server/src/lib/storage.ts
@@ -23,7 +23,8 @@ const schema = {
     default: {
       workId: '',
       name: '请登录',
-      avatarUrl: 'https://img.alicdn.com/tfs/TB1hjBJXLxj_uVjSZFqXXaboFXa-147-150.jpg',
+      avatarUrl:
+        'https://img.alicdn.com/tfs/TB1hjBJXLxj_uVjSZFqXXaboFXa-147-150.jpg',
       isLogin: false,
     },
   },
@@ -47,6 +48,14 @@ const schema = {
     type: 'string',
     default: '@alifd/theme-iceworks-dark',
   },
+  npmClient: {
+    type: 'string',
+    default: '',
+  },
+  registry: {
+    type: 'string',
+    default: 'https://registry.npm.taobao.org',
+  },
   material: {
     type: 'array',
     default: [
@@ -55,6 +64,12 @@ const schema = {
         name: '飞冰物料',
         type: 'react',
         source: 'http://ice.alicdn.com/assets/materials/react-materials.json',
+      },
+      {
+        official: true,
+        name: 'Vue物料',
+        type: 'vue',
+        source: 'http://ice.alicdn.com/assets/materials/vue-materials.json',
       },
     ],
   },
@@ -67,6 +82,8 @@ const schema = {
     default: [],
   },
 };
+
+export { schema };
 
 export default new Conf({
   schema,