+++ tools/iceworks-cli/lib/checkVersion.js
@@ -1,13 +1,10 @@
 const semver = require('semver');
 const { getNpmLatestSemverVersion } = require('ice-npm-utils');
 
-module.exports = async (packageName, packageVersion) => {
-  try {
-    const latestVersion = await getNpmLatestSemverVersion(packageName, packageVersion);
-    if (semver.lt(packageVersion, latestVersion)) {
-      return latestVersion;
+module.exports = (packageName, packageVersion) => {
+  return getNpmLatestSemverVersion(packageName, packageVersion).then(latestVersion => {
+    if (latestVersion && semver.lt(packageVersion, latestVersion)) {
+      return Promise.resolve(latestVersion);
     }
-  } catch (error) {
-    // ...
-  }
+  })
 };